+++ reading-website-biz/src/test/java/com/reading/website/biz/mapper/SmallCategoryMapperTest.java
@@ -20,8 +20,8 @@ public class SmallCategoryMapperTest extends BaseTest {
 
     @Test
     public void testListByBigCateIds() {
-        List<Long> bigCateIds = new ArrayList<>();
-        bigCateIds.add(1L);
+        List<Integer> bigCateIds = new ArrayList<>();
+        bigCateIds.add(1);
         List<SmallCategoryDO> res = mapper.listByBigCateIds(bigCateIds);
         System.out.println(res);
     }
@@ -30,19 +30,19 @@ public class SmallCategoryMapperTest extends BaseTest {
     public void test() {
         List<SmallCategoryDO> list = new ArrayList<>();
 
-        list.add(new SmallCategoryDO("言情", 2L));
-        list.add(new SmallCategoryDO("青春", 2L));
-        list.add(new SmallCategoryDO("都市", 2L));
-        list.add(new SmallCategoryDO("历史", 2L));
-        list.add(new SmallCategoryDO("官场商战", 2L));
-        list.add(new SmallCategoryDO("悬疑推理", 2L));
-        list.add(new SmallCategoryDO("武侠魔幻", 2L));
-        list.add(new SmallCategoryDO("社会乡土", 2L));
-        list.add(new SmallCategoryDO("灵异恐怖", 2L));
-        list.add(new SmallCategoryDO("影视文学", 2L));
-        list.add(new SmallCategoryDO("现当代小说", 2L));
-        list.add(new SmallCategoryDO("外国文学", 2L));
-        list.add(new SmallCategoryDO("其他", 2L));
+        list.add(new SmallCategoryDO("言情", 2));
+        list.add(new SmallCategoryDO("青春", 2));
+        list.add(new SmallCategoryDO("都市", 2));
+        list.add(new SmallCategoryDO("历史", 2));
+        list.add(new SmallCategoryDO("官场商战", 2));
+        list.add(new SmallCategoryDO("悬疑推理", 2));
+        list.add(new SmallCategoryDO("武侠魔幻", 2));
+        list.add(new SmallCategoryDO("社会乡土", 2));
+        list.add(new SmallCategoryDO("灵异恐怖", 2));
+        list.add(new SmallCategoryDO("影视文学", 2));
+        list.add(new SmallCategoryDO("现当代小说", 2));
+        list.add(new SmallCategoryDO("外国文学", 2));
+        list.add(new SmallCategoryDO("其他", 2));
         list.forEach( bigCategoryDO -> {
             int res = mapper.insertSelective(bigCategoryDO);
             System.out.println(res);
